Treat a non-existent dotenv file as an empty one
A dotenv file should complement the environment, that is, variables from
the dotenv are merged with the ENV settings, with a fine control over
a) value overriding and b) changing the environment. This is the current
behavior and it should be preserved.

About the case of a _missing dotenv file_: this should be treated as an
_empty_ dotenv and _not_ raise an error, but produce an outcome equal to
the one of an _empty_ dotenv. This is the change proposed here.